backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.2
dnspython==2.7.0
fastapi==0.116.1
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
motor
# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.0.1
PyJWT==2.8.0
# Google OAuth dependencies
httpx==0.24.1
authlib==1.3.0
routes